worker/requirements.txt
psycopg2-binary
requests
python-dotenv
websocket-client